Moved / renamed Spectrum utils (#1909)
