+++ packages/components/src/index.ts
@@ -49,7 +49,6 @@ export * from './SelectValueList';
 export * from './shortcuts';
 export { default as SocketedButton } from './SocketedButton';
 export * from './spectrum';
-export * from './spectrum/utils';
 export * from './TableViewEmptyState';
 export * from './TextWithTooltip';
 export * from './theme';
