+++ packages/components/src/spectrum/picker/index.ts
@@ -1,2 +1,2 @@
 export * from './Picker';
-export * from './PickerUtils';
+export * from './PickerItemContent';
