+++ packages/components/src/spectrum/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './itemUtils';
+export * from './themeUtils';
