+++ packages/components/src/spectrum/utils/themeUtils.test.ts
@@ -1,4 +1,4 @@
-import { themeDHDefault } from './utils';
+import { themeDHDefault } from './themeUtils';
 
 describe('themeDHDefault', () => {
   it('should merge Spectrum default with DH custom styles', () => {
